# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clab AI - MemeCoin Launch Strategist API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export type UserProfileRiskLevel =
  (typeof UserProfileRiskLevel)[keyof typeof UserProfileRiskLevel];

export const UserProfileRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export type UserProfileCreatorType =
  (typeof UserProfileCreatorType)[keyof typeof UserProfileCreatorType];

export const UserProfileCreatorType = {
  builder: "builder",
  casual: "casual",
  agency: "agency",
} as const;

export interface UserProfile {
  id: number;
  userId: string;
  budget: string;
  riskLevel: UserProfileRiskLevel;
  audience: string;
  timeline: string;
  creatorType: UserProfileCreatorType;
  createdAt: string;
  updatedAt: string;
}

export type CreateProfileBodyRiskLevel =
  (typeof CreateProfileBodyRiskLevel)[keyof typeof CreateProfileBodyRiskLevel];

export const CreateProfileBodyRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export type CreateProfileBodyCreatorType =
  (typeof CreateProfileBodyCreatorType)[keyof typeof CreateProfileBodyCreatorType];

export const CreateProfileBodyCreatorType = {
  builder: "builder",
  casual: "casual",
  agency: "agency",
} as const;

export interface CreateProfileBody {
  budget: string;
  riskLevel: CreateProfileBodyRiskLevel;
  audience: string;
  timeline: string;
  creatorType: CreateProfileBodyCreatorType;
}

export interface MemeProject {
  id: number;
  userId: string;
  idea: string;
  name?: string | null;
  ticker?: string | null;
  narrative?: string | null;
  lore?: string | null;
  roadmap?: string | null;
  brandVoice?: string | null;
  launchThread?: string | null;
  faq?: string | null;
  riskReport?: string | null;
  conversationId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface LaunchScores {
  id: number;
  projectId: number;
  launchPotential: number;
  originality: number;
  memeStrength: number;
  survivability: number;
  riskScore: number;
  communityPotential: number;
  executionDifficulty: number;
  analysis?: string | null;
  createdAt: string;
}

export type ScenarioScenarioType =
  (typeof ScenarioScenarioType)[keyof typeof ScenarioScenarioType];

export const ScenarioScenarioType = {
  viral_growth: "viral_growth",
  slow_organic: "slow_organic",
  failure: "failure",
  copycat_attack: "copycat_attack",
  low_budget: "low_budget",
  hype_collapse: "hype_collapse",
} as const;

export type ScenarioOutputData = { [key: string]: unknown } | null;

export interface Scenario {
  id: number;
  projectId: number;
  scenarioType: ScenarioScenarioType;
  title: string;
  probability: number;
  timeline?: string | null;
  outcome: string;
  keyFactors: string[];
  mitigations: string[];
  outputData?: ScenarioOutputData;
  createdAt: string;
}

export interface MemeProjectDetail {
  id: number;
  userId: string;
  idea: string;
  name?: string | null;
  ticker?: string | null;
  narrative?: string | null;
  lore?: string | null;
  roadmap?: string | null;
  brandVoice?: string | null;
  launchThread?: string | null;
  faq?: string | null;
  riskReport?: string | null;
  conversationId?: number | null;
  scores?: LaunchScores | null;
  scenarios: Scenario[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateProjectBody {
  idea: string;
  name?: string;
  ticker?: string;
}

export interface UpdateProjectBody {
  idea?: string;
  name?: string;
  ticker?: string;
  narrative?: string;
  lore?: string;
  roadmap?: string;
  brandVoice?: string;
  launchThread?: string;
  faq?: string;
  riskReport?: string;
}

export interface SaveScoresBody {
  launchPotential: number;
  originality: number;
  memeStrength: number;
  survivability: number;
  riskScore: number;
  communityPotential: number;
  executionDifficulty: number;
  analysis?: string;
}

export type GenerateContentBodyContentType =
  (typeof GenerateContentBodyContentType)[keyof typeof GenerateContentBodyContentType];

export const GenerateContentBodyContentType = {
  name: "name",
  ticker: "ticker",
  lore: "lore",
  roadmap: "roadmap",
  brandVoice: "brandVoice",
  launchThread: "launchThread",
  faq: "faq",
  riskReport: "riskReport",
  scores: "scores",
} as const;

export interface GenerateContentBody {
  contentType: GenerateContentBodyContentType;
}

export interface GeneratedContent {
  contentType: string;
  content: string;
  scores?: LaunchScores | null;
}

export interface LaunchPlan {
  projectId: number;
  projectName?: string | null;
  ticker?: string | null;
  idea: string;
  narrative?: string | null;
  lore?: string | null;
  roadmap?: string | null;
  brandVoice?: string | null;
  launchThread?: string | null;
  faq?: string | null;
  riskReport?: string | null;
  scores?: LaunchScores | null;
  scenarios: Scenario[];
  exportedAt: string;
}

export type CreateScenarioBodyScenarioType =
  (typeof CreateScenarioBodyScenarioType)[keyof typeof CreateScenarioBodyScenarioType];

export const CreateScenarioBodyScenarioType = {
  viral_growth: "viral_growth",
  slow_organic: "slow_organic",
  failure: "failure",
  copycat_attack: "copycat_attack",
  low_budget: "low_budget",
  hype_collapse: "hype_collapse",
} as const;

export interface CreateScenarioBody {
  projectId: number;
  scenarioType: CreateScenarioBodyScenarioType;
}

export interface DashboardSummary {
  totalProjects: number;
  avgLaunchScore: number;
  totalScenarios: number;
  recentProjects: MemeProject[];
  topProject?: MemeProject | null;
  profileComplete: boolean;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export type GenerateOpenaiImageBodySize =
  (typeof GenerateOpenaiImageBodySize)[keyof typeof GenerateOpenaiImageBodySize];

export const GenerateOpenaiImageBodySize = {
  "1024x1024": "1024x1024",
  "512x512": "512x512",
  "256x256": "256x256",
} as const;

export interface GenerateOpenaiImageBody {
  prompt: string;
  size?: GenerateOpenaiImageBodySize;
}

export interface GenerateOpenaiImageResponse {
  b64_json: string;
}

export interface OpenaiError {
  error: string;
}

export type ListScenariosParams = {
  projectId: number;
};