# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Clab AI - MemeCoin Launch Strategist API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profiles:
    get:
      operationId: getProfile
      tags: [profiles]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: createProfile
      tags: [profiles]
      summary: Create or update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProfileBody"
      responses:
        "201":
          description: Profile created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    put:
      operationId: updateProfile
      tags: [profiles]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpenaiConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"
  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming text response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOpenaiMessageBody"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}
  /openai/generate-image:
    post:
      operationId: generateOpenaiImage
      tags: [openai]
      summary: Generate an image from a text prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateOpenaiImageBody"
      responses:
        "200":
          description: Generated image data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateOpenaiImageResponse"
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        budget:
          type: string
        riskLevel:
          type: string
          enum: [low, medium, high]
        audience:
          type: string
        timeline:
          type: string
        creatorType:
          type: string
          enum: [builder, casual, agency]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - budget
        - riskLevel
        - audience
        - timeline
        - creatorType
        - createdAt
        - updatedAt
    CreateProfileBody:
      type: object
      properties:
        budget:
          type: string
        riskLevel:
          type: string
          enum: [low, medium, high]
        audience:
          type: string
        timeline:
          type: string
        creatorType:
          type: string
          enum: [builder, casual, agency]
      required:
        - budget
        - riskLevel
        - audience
        - timeline
        - creatorType
...
    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    CreateOpenaiConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    SendOpenaiMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - createdAt
        - messages
    GenerateOpenaiImageBody:
      type: object
      properties:
        prompt:
          type: string
        size:
          type: string
          enum: ["1024x1024", "512x512", "256x256"]
      required:
        - prompt
    GenerateOpenaiImageResponse:
      type: object
      properties:
        b64_json:
          type: string
      required:
        - b64_json
    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error