# lib/api-zod/src/generated/types/openaiError.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clab AI - MemeCoin Launch Strategist API
 * OpenAPI spec version: 0.1.0
 */

export interface OpenaiError {
  error: string;
}